Jalsonic Opinion Version v1 Beta
Opinion (o-) is a modern and powerful front-end framework with built-in responsiveness.It Uses standard CSS for Speeds up mobile HTML apps, Removes the inconsistent styling of HTML elements provided by browsers and Provides equality CSS for all devices. PC, laptop, tablet, and mobile. A secure way for developers to enjoy the benefit of a responsiveness, network and geolocation objects and animation websites and mobile web application.